chatbot-backend/requirements.txt
Flask
flask-cors
gunicorn
twilio
langchain
langchain-community
openai
tiktoken
chromadb
beautifulsoup4
markdown
pandas
tqdm
python-dotenv
tenacity
requests
SQLAlchemy
psycopg2-binary
sqlitedict
langdetect
pytest
black
isort
flake8